requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
apprise==1.9.1
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
coolname==2.2.0
croniter==5.0.1
cryptography==44.0.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.1
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
narwhals==1.20.1
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.2.2
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.0
opentelemetry-api==1.29.0
optuna==4.1.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pendulum==3.0.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
postgrest==0.19.1
prefect==3.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.6.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
readchar==4.2.1
realtime==2.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
storage3==0.11.0
streamlit==1.41.1
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==13.1
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
